@@ -3,40 +3,38 @@
   "description": "EEG/iEEG to BIDS format Wizard",
   "version": "1.0.5",
   "dependencies": {
-    "@electron/remote": "^1.1.0",
-    "@fortawesome/fontawesome-free": "^5.15.4",
-    "electron-log": "^4.3.5",
-    "electron-store": "^8.0.0",
-    "keytar": "^7.7.0",
-    "papaparse": "^5.3.1",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.2",
+    "@electron/remote": "^2.0.8",
+    "@fortawesome/fontawesome-free": "^6.1.1",
+    "electron-log": "^4.4.6",
+    "electron-store": "^8.0.1",
+    "keytar": "^7.9.0",
+    "papaparse": "^5.3.2",
+    "prop-types": "^15.8.1",
+    "react": "^18.0.0",
     "react-color": "^2.19.3",
-    "react-datepicker": "^4.1.1",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.2",
-    "react-switch": "^6.0.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.0",
     "react-tooltip": "^4.2.21",
-    "socket.io-client": "^4.1.2"
+    "socket.io-client": "^4.4.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.3",
-    "@babel/eslint-parser": "^7.14.4",
-    "concurrently": "^6.2.0",
+    "@babel/core": "^7.17.9",
+    "@babel/eslint-parser": "^7.17.0",
+    "concurrently": "^7.1.0",
     "cross-env": "^7.0.3",
-    "electron": "^13.0.1",
-    "electron-builder": "^22.11.4",
+    "electron": "^18.0.3",
+    "electron-builder": "^23.0.3",
     "electron-devtools-installer": "^3.2.0",
-    "eslint": "^7.27.0",
+    "eslint": "^8.13.0",
     "eslint-config-google": "^0.14.0",
     "eslint-plugin-babel": "^5.3.1",
-    "eslint-plugin-import": "^2.23.4",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jsx": "^0.1.0",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.24.0",
-    "lint-staged": "^11.0.0",
-    "wait-on": "^6.0.0"
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-react": "^7.29.4",
+    "lint-staged": "^12.3.7",
+    "wait-on": "^6.0.1"
   },
   "scripts": {
     "rebuild": "rebuild --runtime=electron --target=11.2.1",
